Refactor generator to use modules to update styling
