+++ src/pages/generator/index.module.scss
@@ -1,0 +1,20 @@
+.generatedQuestion {
+  font-size: 2.2rem;
+  text-align: center;
+  margin: 3rem 0;
+
+  @media (min-width: 400px) {
+    font-size: 2.4rem;
+    margin: 4rem 0;
+  }
+}
+
+.generatedPassion,
+.generatedPurpose {
+  text-decoration: underline;
+  font-weight: 600;
+}
+
+.generateButtonContainer {
+  margin-top: 1rem;
+}
