+++ src/style/index.scss
@@ -199,25 +199,6 @@ a.link {
   }
 }
 
-.generated-question {
-  font-size: 2.4rem;
-  text-align: center;
-  margin: 3rem 0;
-
-  @media (min-width: 400px) {
-    margin: 4rem 0;
-  }
-  @media (min-width: 600px) {
-    margin: 5rem 0;
-  }
-}
-
-.generated-passion,
-.generated-purpose {
-  text-decoration: underline;
-  font-weight: 600;
-}
-
 .checkbox {
   input {
     opacity: 0;
